address code review comments
