+++ projects/media-viewer/src/lib/store/model/bookmarks.interface.ts
@@ -1,12 +1,3 @@
-
-export interface BookmarksState {
-  bookmarks: Bookmark[],
-  bookmarkEntities: { [id: string]: Bookmark },
-  editableBookmark: string,
-  loaded: boolean,
-  loading: boolean
-}
-
 export interface Bookmark {
   id: string;
   documentId: string;
