+++ projects/media-viewer/src/lib/store/reducers/bookmarks.reducer.ts
@@ -1,6 +1,14 @@
 import * as fromBookmarks from '../actions/bookmarks.action';
-import { StoreUtils } from '../store-utils';
-import { Bookmark, BookmarksState } from '../../viewers/pdf-viewer/side-bar/bookmarks/bookmarks.interfaces';
+import { Bookmark } from '../../viewers/pdf-viewer/side-bar/bookmarks/bookmarks.interfaces';
+import { generateBookmarkEntities } from '../bookmarks-store-utils';
+
+export interface BookmarksState {
+  bookmarks: Bookmark[],
+  bookmarkEntities: { [id: string]: Bookmark },
+  editableBookmark: string,
+  loaded: boolean,
+  loading: boolean
+}
 
 export const initialBookmarksState: BookmarksState = {
   bookmarks: [],
@@ -25,7 +33,7 @@ export function bookmarksReducer (state = initialBookmarksState,
     case fromBookmarks.LOAD_BOOKMARKS_SUCCESS:
     case fromBookmarks.LOAD_BOOKMARKS_FAIL:{
       const bookmarks = action.payload.status === 200 ? action.payload.body : [];
-      const bookmarkEntities = StoreUtils.generateBookmarkEntities(bookmarks);
+      const bookmarkEntities = generateBookmarkEntities(bookmarks);
       return {
         ...state,
         bookmarks,
@@ -51,7 +59,7 @@ export function bookmarksReducer (state = initialBookmarksState,
     }
 
     case fromBookmarks.MOVE_BOOKMARK_SUCCESS: {
-      const movedBookmarks = StoreUtils.generateBookmarkEntities(action.payload);
+      const movedBookmarks = generateBookmarkEntities(action.payload);
       const bookmarkEntities = {
         ...state.bookmarkEntities,
         ...movedBookmarks
