+++ projects/media-viewer/src/lib/store/store-utils.ts
@@ -1,5 +1,4 @@
-import {Annotation} from '../annotations/annotation-set/annotation-view/annotation.model';
-import { Bookmark } from '../viewers/pdf-viewer/side-bar/bookmarks/bookmarks.interfaces';
+import { Annotation } from '../annotations/annotation-set/annotation-view/annotation.model';
 import * as moment_ from 'moment-timezone';
 
 /*
@@ -98,13 +97,6 @@ export class StoreUtils {
     }, {});
   }
 
-  static generateBookmarkEntities(bookmarks: Bookmark[]): {[id: string]: Bookmark} {
-    return bookmarks.reduce(
-      (bookmarkEntities, bookmark) =>
-        Object.assign(bookmarkEntities, { [bookmark.id]: bookmark }),
-      {});
-  }
-
   static filterCommentsSummary(comments, filters) {
     if(Object.keys(filters).length) {
       const tagFilterApplied = Object.keys(filters.tagFilters)
@@ -153,54 +145,6 @@ export class StoreUtils {
       return (tagFilterApplied || dateFilterApplied) ? filteredComments : comments;
     } else {
       return comments;
-    }
-  }
-
-  static generateBookmarkNodes(entities: { [id: string]: Bookmark }) {
-    const bookmarkEntities = JSON.parse(JSON.stringify(entities));
-    this.indexEntities(bookmarkEntities);
-    return Object.keys(bookmarkEntities).reduce((nodes, bookmarkId) => {
-      const bookmarkEntity = bookmarkEntities[bookmarkId];
-      if (bookmarkEntity.parent) {
-        const parentEntity = bookmarkEntities[bookmarkEntity.parent];
-        if (!parentEntity.children) {
-          parentEntity.children = [];
-        }
-        parentEntity.children[bookmarkEntity.index] = bookmarkEntity;
-      } else {
-        nodes[bookmarkEntity.index] = bookmarkEntity;
-      }
-      return nodes;
-    }, []);
-  }
-
-  static indexEntities(entities: { [id: string]: Bookmark }) {
-    const entityIds = Object.keys(entities);
-    let index = 0;
-    while (entityIds.length > 0) {
-      let keysToRemove = [];
-      entityIds.forEach(key => {
-        if (!entities[key].previous || !entityIds.includes(entities[key].previous.toString())) {
-          entities[key].index = index;
-          keysToRemove.push(key);
-        }
-      });
-      keysToRemove.forEach(key => entityIds.splice(entityIds.indexOf(key), 1));
-      keysToRemove = [];
-      index++;
-    }
-  }
-
-  static getAllChildren(bookmarks:Bookmark[]) {
-    if (bookmarks) {
-      return bookmarks.reduce((childIds, bookmark) => {
-        if (bookmark.children && bookmark.children.length > 0) {
-          return [...childIds, bookmark.id, ...this.getAllChildren(bookmark.children)]
-        }
-        return [...childIds, bookmark.id]
-      }, []);
-    } else {
-      return []
     }
   }
 }
