+++ projects/media-viewer/src/lib/viewers/pdf-viewer/side-bar/bookmarks/bookmarks.interfaces.ts
@@ -1,11 +1,3 @@
-export interface BookmarksState {
-  bookmarks: Bookmark[],
-  bookmarkEntities: { [id: string]: Bookmark },
-  editableBookmark: string,
-  loaded: boolean,
-  loading: boolean
-}
-
 export interface Bookmark {
   id: string;
   documentId: string;
